+++ pyanalyze/test_typeis.py
@@ -126,10 +126,8 @@ class TestTypeIs(TestNameCheckVisitorBase):
 
     @assert_passes()
     def testTypeIsHigherOrder(self):
-        import collections.abc
         from typing import Callable, TypeVar, Iterable, List
-        from typing_extensions import TypeIs
-        from pyanalyze.value import assert_is_value, GenericValue, AnyValue, AnySource
+        from typing_extensions import TypeIs, assert_type
 
         T = TypeVar("T")
         R = TypeVar("R")
@@ -143,13 +141,7 @@ class TestTypeIs(TestNameCheckVisitorBase):
         def capybara() -> None:
             a: List[object] = ["a", 0, 0.0]
             b = filter(is_float, a)
-            # TODO should be Iterable[float]
-            assert_is_value(
-                b,
-                GenericValue(
-                    collections.abc.Iterable, [AnyValue(AnySource.generic_argument)]
-                ),
-            )
+            assert_type(b, Iterable[float])
 
     @assert_passes()
     def testTypeIsMethod(self):
@@ -277,10 +269,8 @@ class TestTypeIs(TestNameCheckVisitorBase):
 
     @assert_passes()
     def testTypeIsOverload(self):
-        import collections.abc
         from typing import Callable, Iterable, Iterator, List, Optional, TypeVar
-        from typing_extensions import TypeIs, overload
-        from pyanalyze.value import assert_is_value, GenericValue, AnyValue, AnySource
+        from typing_extensions import TypeIs, overload, assert_type
 
         T = TypeVar("T")
         R = TypeVar("R")
@@ -302,21 +292,13 @@ class TestTypeIs(TestNameCheckVisitorBase):
         def is_int_bool(a: object) -> bool:
             return False
 
-        iter_any = GenericValue(
-            collections.abc.Iterator, [AnyValue(AnySource.generic_argument)]
-        )
-
         def main(a: List[Optional[int]]) -> None:
             bb = filter(lambda x: x is not None, a)
-            # TODO Iterator[Optional[int]]
-            assert_is_value(bb, iter_any)
-            # Also, if you replace 'bool' with 'Any' in the second overload, bb is Iterator[Any]
+            assert_type(bb, Iterator[Optional[int]])
             cc = filter(is_int_typeguard, a)
-            # TODO Iterator[int]
-            assert_is_value(cc, iter_any)
+            assert_type(cc, Iterator[int])
             dd = filter(is_int_bool, a)
-            # TODO Iterator[Optional[int]]
-            assert_is_value(dd, iter_any)
+            assert_type(dd, Iterator[Optional[int]])
 
     @assert_passes()
     def testTypeIsDecorated(self):
@@ -345,7 +327,7 @@ class TestTypeIs(TestNameCheckVisitorBase):
                 return False
 
         class D(C):
-            def is_float(self, a: object) -> bool:  # TODO: incompatible_override
+            def is_float(self, a: object) -> bool:  # E: incompatible_override
                 return False
 
     @assert_passes()
@@ -576,10 +558,10 @@ class TestTypeIs(TestNameCheckVisitorBase):
             return False
 
         accepts_typeguard(with_typeguard)
-        accepts_typeguard(with_bool)  # TODO error
+        accepts_typeguard(with_bool)  # E: incompatible_argument
 
-        different_typeguard(with_typeguard)  # TODO error
-        different_typeguard(with_bool)  # TODO error
+        different_typeguard(with_typeguard)  # E: incompatible_argument
+        different_typeguard(with_bool)  # E: incompatible_argument
 
     @assert_passes()
     def testTypeIsAsGenericFunctionArg(self):
@@ -602,7 +584,7 @@ class TestTypeIs(TestNameCheckVisitorBase):
 
         accepts_typeguard(with_bool_typeguard)
         accepts_typeguard(with_str_typeguard)
-        accepts_typeguard(with_bool)  # TODO error
+        accepts_typeguard(with_bool)  # E: incompatible_argument
 
     @assert_passes()
     def testTypeIsAsOverloadedFunctionArg(self):
@@ -662,9 +644,9 @@ class TestTypeIs(TestNameCheckVisitorBase):
         def with_typeguard_c(o: object) -> TypeIs[C]:
             return False
 
-        accepts_typeguard(with_typeguard_a)  # TODO error
+        accepts_typeguard(with_typeguard_a)  # E: incompatible_argument
         accepts_typeguard(with_typeguard_b)
-        accepts_typeguard(with_typeguard_c)  # TODO error
+        accepts_typeguard(with_typeguard_c)  # E: incompatible_argument
 
     @assert_passes()
     def testTypeIsWithIdentityGeneric(self):
@@ -786,7 +768,7 @@ class TestTypeIs(TestNameCheckVisitorBase):
         @overload
         def typeguard(x: object, y: int) -> TypeIs[int]: ...
 
-        def typeguard(x: object, y: Union[int, str]) -> Union[TypeIs[int], TypeIs[str]]:
+        def typeguard(x: object, y: Union[int, str]) -> bool:
             return False
 
         def capybara(x: object) -> None:
@@ -805,7 +787,7 @@ class TestTypeIs(TestNameCheckVisitorBase):
     @assert_passes()
     def testGenericAliasWithTypeIs(self):
         from typing import Callable, List, TypeVar
-        from typing_extensions import TypeIs
+        from typing_extensions import TypeIs, assert_type
 
         T = TypeVar("T")
         A = Callable[[object], TypeIs[List[T]]]
@@ -817,8 +799,7 @@ class TestTypeIs(TestNameCheckVisitorBase):
             raise NotImplementedError
 
         def capybara() -> None:
-            pass
-            # TODO: assert_type(test(foo), List[str])
+            assert_type(test(foo), str)
 
     @assert_passes()
     def testNoCrashOnDunderCallTypeIs(self):